requirements.txt
selenium==4.15.0
playwright==1.40.0
beautifulsoup4==4.12.0
tensorflow>=2.16.0
pandas==2.1.0
opencv-python==4.8.0
pillow==10.0.0
requests==2.31.0
schedule==1.2.0
python-dotenv==1.0.0
cryptography==41.0.0
openpyxl==3.1.0
lxml==4.9.0
numpy==1.24.0
matplotlib==3.7.0
psutil==5.9.0